requirements.txt
yt-dlp>=2023.12.30
openai-whisper>=20231117
click>=8.1.7
pydub>=0.25.1
ffmpeg-python>=0.2.0
tqdm>=4.66.1
llm>=0.13.0
llm-gemini>=0.1.0